+++ src/config.js
@@ -54,8 +54,8 @@ class Config {
   #env;
   #eventMap;
   #updatePeriodicEvents;
-  #blockedPeriodicEvents;
-  #isPeriodicEventBlockedCb;
+  #blockedEvents;
+  #isEventBlockedCb;
   #thresholdLoggingEventProcessing;
   #useAsCAPOutbox;
   #userId;
@@ -79,7 +79,7 @@ class Config {
     this.#skipCsnCheck = null;
     this.#disableRedis = null;
     this.#env = getEnvInstance();
-    this.#blockedPeriodicEvents = {};
+    this.#blockedEvents = {};
   }
 
   getEventConfig(type, subType) {
@@ -103,7 +103,7 @@ class Config {
   }
 
   attachConfigChangeHandler() {
-    this.#attachBlocklistChangeHandler();
+    this.#attachBlockListChangeHandler();
     redis.subscribeRedisChannel(REDIS_CONFIG_CHANNEL, (messageData) => {
       try {
         const { key, value } = JSON.parse(messageData);
@@ -129,14 +129,14 @@ class Config {
     });
   }
 
-  #attachBlocklistChangeHandler() {
+  #attachBlockListChangeHandler() {
     redis.subscribeRedisChannel(REDIS_CONFIG_BLOCKLIST_CHANNEL, (messageData) => {
       try {
         const { command, key, tenant } = JSON.parse(messageData);
         if (command === COMMAND_BLOCK) {
-          this.#blockPeriodicEventLocalState(key, tenant);
+          this.#blockEventLocalState(key, tenant);
         } else {
-          this.#unblockPeriodicEventLocalState(key, tenant);
+          this.#unblockEventLocalState(key, tenant);
         }
       } catch (err) {
         this.#logger.error("could not parse event blocklist change", err, {
@@ -146,14 +146,14 @@ class Config {
     });
   }
 
-  blockPeriodicEvent(type, subType, tenant = "*") {
-    const typeWithSuffix = `${type}${SUFFIX_PERIODIC}`;
+  blockEvent(type, subType, isPeriodic, tenant = "*") {
+    const typeWithSuffix = `${type}${isPeriodic ? SUFFIX_PERIODIC : ""}`;
     const config = this.getEventConfig(typeWithSuffix, subType);
     if (!config) {
       return;
     }
     const key = this.generateKey(typeWithSuffix, subType);
-    this.#blockPeriodicEventLocalState(key, tenant);
+    this.#blockEventLocalState(key, tenant);
     if (!this.redisEnabled) {
       return;
     }
@@ -165,24 +165,24 @@ class Config {
       });
   }
 
-  #blockPeriodicEventLocalState(key, tenant) {
-    this.#blockedPeriodicEvents[key] ??= {};
-    this.#blockedPeriodicEvents[key][tenant] = true;
+  #blockEventLocalState(key, tenant) {
+    this.#blockedEvents[key] ??= {};
+    this.#blockedEvents[key][tenant] = true;
     return key;
   }
 
   clearPeriodicEventBlockList() {
-    this.#blockedPeriodicEvents = {};
+    this.#blockedEvents = {};
   }
 
-  unblockPeriodicEvent(type, subType, tenant = "*") {
-    const typeWithSuffix = `${type}${SUFFIX_PERIODIC}`;
+  unblockEvent(type, subType, isPeriodic, tenant = "*") {
+    const typeWithSuffix = `${type}${isPeriodic ? SUFFIX_PERIODIC : ""}`;
     const key = this.generateKey(typeWithSuffix, subType);
     const config = this.getEventConfig(typeWithSuffix, subType);
     if (!config) {
       return;
     }
-    this.#unblockPeriodicEventLocalState(key, tenant);
+    this.#unblockEventLocalState(key, tenant);
     if (!this.redisEnabled) {
       return;
     }
@@ -218,17 +218,17 @@ class Config {
     this.#eventMap[this.generateKey(CAP_EVENT_TYPE, serviceName)] = eventConfig;
   }
 
-  #unblockPeriodicEventLocalState(key, tenant) {
-    const map = this.#blockedPeriodicEvents[key];
+  #unblockEventLocalState(key, tenant) {
+    const map = this.#blockedEvents[key];
     if (!map) {
       return;
     }
-    this.#blockedPeriodicEvents[key][tenant] = false;
+    this.#blockedEvents[key][tenant] = false;
     return key;
   }
 
-  isPeriodicEventBlocked(type, subType, tenant) {
-    const map = this.#blockedPeriodicEvents[this.generateKey(type, subType)];
+  isEventBlocked(type, subType, tenant) {
+    const map = this.#blockedEvents[this.generateKey(type, subType)];
     if (!map) {
       return false;
     }
@@ -376,12 +376,12 @@ class Config {
     this.#instanceLoadLimit = value;
   }
 
-  get isPeriodicEventBlockedCb() {
-    return this.#isPeriodicEventBlockedCb;
+  get isEventBlockedCb() {
+    return this.#isEventBlockedCb;
   }
 
-  set isPeriodicEventBlockedCb(value) {
-    this.#isPeriodicEventBlockedCb = value;
+  set isEventBlockedCb(value) {
+    this.#isEventBlockedCb = value;
   }
 
   get tableNameEventQueue() {